fix: adding module props in src subgen

+++ src/src/src.ts
@@ -1,4 +1,6 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
+import sort from 'sort-package-json';
+// eslint-disable-next-line import/no-extraneous-dependencies
 import type { PackageJson } from 'type-fest';
 import Generator from 'yeoman-generator';
 import { configs, dependencies } from '../core';
@@ -10,11 +12,7 @@ export default class SrcGenerator extends Generator {
 
   public async writing(): Promise<void> {
     const pkg = (await configs.load('package.json')) as PackageJson;
-    const { main, name } = pkg;
-
-    if (typeof main !== 'undefined' && this.fs.exists(this.destinationPath(main))) {
-      return;
-    }
+    const { name } = pkg;
 
     let extension = 'js';
 
@@ -41,12 +39,19 @@ export default class SrcGenerator extends Generator {
       packageProps.typings = `${name as string}.d.ts`;
     }
 
-    configs.set('package.json', {
-      ...pkg,
-      ...packageProps,
-    });
+    configs.set(
+      'package.json',
+      sort({
+        ...pkg,
+        ...packageProps,
+      }),
+    );
 
     const mainFile = `src/${name as string}.${extension}`;
-    this.fs.write(this.destinationPath(mainFile), '');
+    if (!this.fs.exists(this.destinationPath(mainFile))) {
+      this.fs.write(this.destinationPath(mainFile), '');
+    }
+
+    await configs.saveAll();
   }
 }
